requirements.txt
fastapi
uvicorn
openai
requests
beautifulsoup4
selenium
webdriver-manager
pytesseract
Pillow
python-multipart